requirements.txt
aiogram==2.21
asyncpg==0.25.0
python-dotenv==0.20.0